+++ leia-refresher/src/main/java/com/grookage/leia/provider/refresher/LeiaRefresher.java
@@ -22,4 +22,5 @@ public interface LeiaRefresher<T> {
 
     void refresh();
 
+    void start();
 }